+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   It was already required in v1.22.2.
 ### Migrations and checks
 #### Schema migrations
+- [#165](https://github.com/LayerManager/layman/issues/165) Add column `role_name` to table `rights` in prime DB schema. Add constraint that exactly one of columns `role_name` and `id_user` is not null.
 #### Data migrations
 ### Changes
 - All changes from [v1.22.1](#v1221) and [v1.22.2](#v1222).